Mise en place du CRUD simple Elimination

+++ src/main/java/com/dreams/gestiontournois/repository/SimpleEliminationRepository.java
@@ -3,5 +3,5 @@ package com.dreams.gestiontournois.repository;
 import com.dreams.gestiontournois.model.SimpleElimination;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SimpleEliminationRepository extends JpaRepository<SimpleElimination, Integer> {
+public interface SimpleEliminationRepository extends JpaRepository<SimpleElimination, Long> {
 }
